# back_end/school-app/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API SCHOOL",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://school-app.test/api"
        }
    ],
    "paths": {
...
        "/admin/student": {
            "get": {
                "tags": [
                    "student"
                ],
                "summary": "listar",
                "description": "Listar studiantes",
                "operationId": "getStudent",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Internal Server"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/student/register": {
            "post": {
                "tags": [
                    "student"
                ],
                "summary": "Register Student",
                "description": "Registrar al Estudiante",
                "operationId": "getStudentRegister",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Nombre",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date_start",
                        "in": "query",
                        "description": "fecha de inicio",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_end",
                        "in": "query",
                        "description": "fecha de fin",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "schedule",
                        "in": "query",
                        "description": "horario",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "tipo de modalidad",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/admin/student/{id}": {
            "get": {
                "tags": [
                    "student"
                ],
                "summary": "obtener",
                "description": "Obtener Estudiante",
                "operationId": "getStudentId",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "el id del Estudiante",
                        "required": true,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "204": {
                        "description": "No Content"
                    }
                }
            },
            "delete": {
                "tags": [
                    "student"
                ],
                "summary": "Eliminar",
                "description": "Eliminar Estudiante",
                "operationId": "getStudentDeleteId",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "el id del Estudiante",
                        "required": true,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/admin/student/updated/{id}": {
            "post": {
                "tags": [
                    "student"
                ],
                "summary": "updated student",
                "description": "Actualizar al Estudiante",
                "operationId": "getStudentUpdated",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Nombre",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date_start",
                        "in": "query",
                        "description": "fecha de inicio",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "date_end",
                        "in": "query",
                        "description": "fecha de fin",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "schedule",
                        "in": "query",
                        "description": "horario",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "tipo de modalidad",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Iinternal Serve"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "Bearer"
            }
        }
    }
}